requirements.txt
# Core
Django==4.2.16
whitenoise==6.7.0
gunicorn==22.0.0

# AI
groq==0.11.0

# Supabase + JWT
supabase==2.9.1
python-jose[cryptography]==3.3.0

# File processing
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
Pillow==10.4.0

# Web search
duckduckgo-search==6.3.7
beautifulsoup4==4.12.3

# Utilities
requests==2.32.3
python-dotenv==1.0.1
